서류관리 완료, 출퇴근 노선관리 아래쪽 리스트 완료

+++ document/views.py
@@ -16,20 +16,24 @@ FORMAT = "%Y-%m-%d"
 
 class CompanyDocumentList(generic.ListView):
     template_name = 'document/company.html'
-    context_object_name = 'file_list'
-    model = Document
+    context_object_name = 'group_list'
+    model = DocumentGroup
 
-    def get_queryset(self):
-        filename = self.request.GET.get('filename')
-        if filename:
-            file_list = Document.objects.filter(filename__contains=filename)
-            return file_list
-        else:
-            return super().get_queryset()
-    
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context['group_list'] = DocumentGroup.objects.all()
+        file_list = ''
+        filename = self.request.GET.get('filename', '')
+        if filename:
+            group_list = []
+            file_list = Document.objects.select_related('group_id').filter(filename__contains=filename)
+            for file in file_list:
+                group_list.append(file.group_id)
+            
+            group_list = list(set(group_list))
+            context['group_list'] = group_list
+
+        context['search'] = filename
+        context['file_list'] = file_list
         return context
 
 def company_document_create(request):
@@ -71,8 +75,7 @@ def company_document_download(request, id):
             quote_file_url = urllib.parse.quote(download_file.filename.encode('utf-8'))
             response = HttpResponse(fh.read(), content_type=mimetypes.guess_type(url)[0])
             response['Content-Disposition'] = 'attachment;filename*=UTF-8\'\'%s' % quote_file_url
-            # return response
-            return JsonResponse({'url': quote_file_url})
+            return response
         raise Http404
     else:
         print("에러")
@@ -84,24 +87,34 @@ def company_delete(request):
 
     for file_id in file_list:
         file = get_object_or_404(Document, id=file_id)
-        os.remove(file.file.path)
+        if file.file:
+            os.remove(file.file.path)
         file.delete()
 
     return redirect(request.META.get('HTTP_REFERER', 'redirect_if_referer_not_found'))
 
 def company_group_delete(request):
-    group_list = request.POST.getlist('group_check')
-
-    for group_id in group_list:
-        group = get_object_or_404(DocumentGroup, id=group_id)
-        document_list = group.group_file.all()
-        for document in document_list:
-            os.remove(document.file.path)
-            document.delete()
-        group.delete()
+    
+    id = request.POST.get('id')
+    group = get_object_or_404(DocumentGroup, id=id)
+    document_list = group.group_file.all()
+    for document in document_list:
+        os.remove(document.file.path)
+        document.delete()
+    group.delete()
     
     return redirect(request.META.get('HTTP_REFERER', 'redirect_if_referer_not_found'))
 
+def company_group_edit(request):
+    name = request.POST.get('name')
+    id = request.POST.get('id')
+
+    group = get_object_or_404(DocumentGroup, id=id)
+    group.name = name
+    group.save()
+
+
+    return redirect(request.META.get('HTTP_REFERER', 'redirect_if_referer_not_found'))
 # def download(request, kinds, notice_id, file_id):
 #     kinds_check(kinds)
 #     download_file = get_object_or_404(NoticeFile, pk=file_id)
@@ -131,52 +144,25 @@ class DocumentList(generic.ListView):
     def get_queryset(self):
         date1 = self.request.GET.get('date1', TODAY)
         date2 = self.request.GET.get('date2', TODAY)
+        route = self.request.GET.get('route')
+        
         order_list = DispatchOrder.objects.prefetch_related('info_order').filter(arrival_date__lte=f'{date2} 24:00').filter(departure_date__gte=f'{date1} 00:00')
+        if route:
+            order_list = order_list.filter(route__contains=route)
         return order_list
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        departure_date = []
-        arrival_date = []
-        time = []
-        num_days = []
 
-        for order in context['order_list']:
-            d_y = order.departure_date[0:4]
-            d_m = order.departure_date[5:7]
-            d_d = order.departure_date[8:10]
-            d_t = order.departure_date[11:16]
-            d_date = date(int(d_y), int(d_m), int(d_d))
-            
-            d_y = d_y[2:4]
-
-            a_y = order.arrival_date[0:4]
-            a_m = order.arrival_date[5:7]
-            a_d = order.arrival_date[8:10]
-            a_t = order.arrival_date[11:16]
-            a_date = date(int(a_y), int(a_m), int(a_d))
-            
-            a_y = a_y[2:4]
-            
-            date_diff = (a_date - d_date) + timedelta(days=1)
-            if date_diff.days > 1:
-                num_days.append(date_diff.days)
-            else:
-                num_days.append('')
-
-            departure_date.append(f"{d_y}.{d_m}.{d_d}")
-            time.append(f"{d_t}~{a_t}")
-            arrival_date.append(f"{a_y}.{a_m}.{a_d}")
-
+        context['route'] = self.request.GET.get('route', '')
+        
+        
         connect_list = []
         for order in context['order_list']:
             connect_list.append(order.info_order.all())
         print("CONNETCTT", connect_list)
         context['connect_list'] = connect_list
-        context['departure_date'] = departure_date
-        context['arrival_date'] = arrival_date
-        context['num_days'] = num_days
-        context['time'] = time
+        
         context['date1'] = self.request.GET.get('date1', TODAY)
         context['date2'] = self.request.GET.get('date2', TODAY)
         
@@ -206,7 +192,7 @@ def commitment_print(request):
     if len(connect_list_all) > 18:
         cut = True
         connect_list = connect_list_all[:18]
-        connect_list2 = connect_list_all[19:]
+        connect_list2 = connect_list_all[18:]
     
     return render(request, 'document/commitment_print.html', {'connect_list': connect_list, 'connect_list2': connect_list2, 'order': order, 'cut': cut})
